requirements.txt
flask
tensorflow
numpy
opencv-python
kafka-python
pyspark
mlflow
pyngrok